% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiometry.R
\docType{package}
\name{audiometry}
\alias{audiometry}
\alias{audiometry-package}
\title{audiometry}
\description{
a package for standard confirm pur tone audiomety data visualisation
using the versatile \code{ggplot2} package.
}
\details{
Right now this is almost only about the \code{gg_pta} function to start a
ggplot with pure tone audiometry data so that the reference frame looks
familiar to audiologists and ent doctors.
}
